+++ libs/cardano-ledger-core/testlib/Test/Cardano/Ledger/Core/Arbitrary.hs
@@ -40,7 +40,6 @@ import Cardano.Ledger.BaseTypes (
   BlocksMade (..),
   CertIx (..),
   DnsName,
-  EpochInterval (..),
   Exclusive (..),
   HasZero,
   Inclusive (..),
@@ -266,9 +265,6 @@ instance Arbitrary Nonce where
       [ return NeutralNonce
       , mkNonceFromNumber <$> arbitrary
       ]
-
-instance Arbitrary EpochInterval where
-  arbitrary = EpochInterval <$> arbitrary
 
 ------------------------------------------------------------------------------------------
 -- Cardano.Ledger.TxIn -------------------------------------------------------------------